move `preprocess/validate` and `postprocess/report` to `utils`
